+++ src/Native/Unix/System.Native/pal_io.h
@@ -697,4 +697,4 @@ extern "C" int32_t SystemNative_GetPeerID(intptr_t socket, uid_t* euid);
 *
 * Returns 0 on success, or -1 if an error occurred (in which case, errno is set appropriately).
 */
-extern "C" int32_t SystemNative_LockFileRegion(int fd, int64_t offset, int64_t length, int16_t lockType);
+extern "C" int32_t SystemNative_LockFileRegion(intptr_t fd, int64_t offset, int64_t length, int16_t lockType);
